python/requirements.txt
csscompressor==0.9.5
htmlmin4==0.2.15
